refactor: move tools directory into agents
Relocates src/lib/tools/ to src/lib/agents/tools/ to co-locate tool
implementations with the agent roles that define their behaviour.
