+++ src/lib/agents/tools/registries.ts
